- bug in plugin_handler->_after_delete - bugs in plugin_automenu
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@1337 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

+++ sys/flexyadmin/controllers/admin/edit.php
@@ -92,16 +92,16 @@ class Edit extends AdminController {
 						$this->db->where(PRIMARY_KEY,$id);
 						$oldData=$this->db->get_row($table);
 
-						$this->_after_delete($table,$oldData);
-						$this->crud->table($table)->delete( array(PRIMARY_KEY=>$id) );
-						
-						/**
-						 * End messages
-						 */
-						$this->load->model("login_log");
-						$this->login_log->update($table);
+						if ($this->_after_delete($table,$oldData)) {
+							$this->crud->table($table)->delete( array(PRIMARY_KEY=>$id) );	
 
-						$this->set_message(langp("delete_succes",$table) . $message);
+							/**
+							 * End messages
+							 */
+							$this->load->model("login_log");
+							$this->login_log->update($table);
+							$this->set_message(langp("delete_succes",$table) . $message);
+						}
 					}
 				}
 				else {
